Move type-related API to the correct header file.
